Move base streamer classes to streamers namespace
