+++ include/web_video_server/streamers/jpeg_streamers.hpp
@@ -40,9 +40,9 @@
 #include "async_web_server_cpp/http_connection.hpp"
 #include "rclcpp/node.hpp"
 
-#include "web_video_server/image_transport_streamer.hpp"
 #include "web_video_server/multipart_stream.hpp"
 #include "web_video_server/streamer.hpp"
+#include "web_video_server/streamers/image_transport_streamer.hpp"
 
 namespace web_video_server
 {
