+++ include/web_video_server/streamers/vp8_streamer.hpp
@@ -37,8 +37,8 @@
 #include "async_web_server_cpp/http_request.hpp"
 #include "rclcpp/node.hpp"
 
-#include "web_video_server/libav_streamer.hpp"
 #include "web_video_server/streamer.hpp"
+#include "web_video_server/streamers/libav_streamer.hpp"
 
 namespace web_video_server
 {
