+++ src/streamers/jpeg_streamers.cpp
@@ -48,8 +48,8 @@
 #include "async_web_server_cpp/http_request.hpp"
 #include "rclcpp/node.hpp"
 
-#include "web_video_server/image_transport_streamer.hpp"
 #include "web_video_server/streamer.hpp"
+#include "web_video_server/streamers/image_transport_streamer.hpp"
 
 namespace web_video_server
 {
